backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
aiosqlite
alembic
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
redis
celery[redis]
httpx
pytest
pytest-asyncio
pytest-httpx
email-validator
bcrypt
